Enable access to Windows Local Machine store.
This commit adds a new method OpenStoreAdvanced which takes a single
bool paramater which indicates if the local machine/system store
should be opened. This paramater is a noop on macOS. On Windows
it causes the CERT_SYSTEM_STORE_LOCAL_MACHINE store to be opened instead
of the user store.